+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # Unreleased
+
+# 0.12.0 - 2022-08-20
 * Update egui to `0.19.0`
 * You can now scale the GUI with `egui_ctx.set_pixels_per_point(…)`.
 * Add miniquad context as a parameter in the lambda in `EguiMq::run`